backend/requirements.txt
flask
flask-cors
google-cloud-bigquery
requests
gunicorn
elasticsearch